# swagger.yaml
openapi: 3.1.0
info:
  title: E-Market API
  description: API pour la gestion d'une plateforme e-commerce
  version: 1.0.0
  contact:
    name: E-Market Team
    email: support@emarket.com
servers:
  - url: http://localhost:3000
    description: Serveur de développement
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Entrez votre token JWT (sans 'Bearer ')"
  schemas:
    Product:
      type: object
      properties:
        _id:
          type: string
          example: '68e7b5010bf901dfd66d774a'
        title:
          type: string
          example: 'Casque Bluetooth Sony WH-1000XM5'
        description:
          type: string
          example: 'Casque antibruit haut de gamme'
        price:
          type: number
          example: 3599
        stock:
          type: integer
          example: 12
        category:
          type: string
          example: 'Électronique'
        images:
          type: array
          items:
            type: string
          example:
            [
              '/images/products/product-1234567890.jpg',
              '/images/products/product-1234567891.jpg',
            ]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Review:
      type: object
      properties:
        _id:
          type: string
          example: '68e7b5010bf901dfd66d774a'
        user:
          type: string
          example: '68e7b5010bf901dfd66d774a'
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: 'Excellent produit!'
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: Liste tous les produits
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des produits
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '401':
          description: Token manquant ou invalide
  /api/products/create:
    post:
      tags: [Products]
      summary: Crée un nouveau produit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - description
                - prix
                - stock
                - categories
              properties:
                title:
                  type: string
                  example: 'Casque Bluetooth Sony WH-1000XM5'
                description:
                  type: string
                  example: 'Casque antibruit haut de gamme avec autonomie de 30h.'
                prix:
                  type: number
                  example: 3599
                stock:
                  type: integer
                  example: 12
                categories:
                  type: string
                  example: '["categoryId1", "categoryId2"]'
                  description: 'JSON array of category IDs'
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: 'Product images (max 5, formats: jpg, jpeg, png, gif, max 5MB each)'
      responses:
        '201':
          description: Produit créé avec succès
        '400':
          description: Données invalides
        '401':
          description: Token manquant ou invalide
  /api/products/update/{id}:
    put:
      tags: [Products]
      summary: Met à jour un produit
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: '68e7b5010bf901dfd66d774a'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                prix:
                  type: number
                stock:
                  type: integer
                categories:
                  type: string
                  example: '["categoryId1", "categoryId2"]'
                  description: 'JSON array of category IDs'
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: 'Product images (max 5, formats: jpg, jpeg, png, gif, max 5MB each)'
      responses:
        '200':
          description: Produit mis à jour
        '400':
          description: Données invalides
        '401':
          description: Token manquant ou invalide
        '404':
          description: Produit non trouvé
  /api/products/delete/{id}:
    delete:
      tags: [Products]
      summary: Supprime un produit
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: '68e7b5010bf901dfd66d774a'
      responses:
        '200':
          description: Produit supprimé
        '401':
          description: Token manquant ou invalide
        '404':
          description: Produit non trouvé
  /api/products/search:
    get:
      tags: [Products]
      summary: Recherche de produits
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filtrer par catégorie
        - in: query
          name: minPrice
          schema:
            type: number
          description: Prix minimum
        - in: query
          name: maxPrice
          schema:
            type: number
          description: Prix maximum
        - in: query
          name: name
          schema:
            type: string
          description: Rechercher par nom
      responses:
        '200':
          description: Résultats de recherche
  /api/products/{productId}/reviews:
    post:
      tags: [Products]
      summary: Ajouter un avis sur un produit
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: '68e7b5010bf901dfd66d774a'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rating
                - comment
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
                comment:
                  type: string
                  example: 'Excellent produit, très satisfait!'
      responses:
        '201':
          description: Avis ajouté
        '400':
          description: Données invalides ou avis déjà existant
        '401':
          description: Non authentifié
        '404':
          description: Produit non trouvé
    get:
      tags: [Products]
      summary: Récupérer les avis d'un produit
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: '68e7b5010bf901dfd66d774a'
      responses:
        '200':
          description: Liste des avis
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
        '404':
          description: Produit non trouvé
...
  /api/seller/products:
    get:
      tags: [Seller Management]
      summary: Récupérer les produits du seller connecté
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Numéro de page
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Nombre d'éléments par page
      responses:
        '200':
          description: Liste des produits du seller
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    type: object
                    properties:
                      currentPage:
                        type: integer
                        example: 1
                      totalPages:
                        type: integer
                        example: 5
                      totalProducts:
                        type: integer
                        example: 47
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé - rôle seller requis
...